Move camel-core-osgi-activator tests to Karaf i...
...tests

Since OSGI activator depends on Karaf features it can only be built
after it, and since Karaf features include OSGI activator it fails as
the activator is not present in the Maven reactor at that point -- it
will be built right after it. This moves the tests, the only part that
depends on the Karaf features to the Karaf integration tests to prevent
this situation.
