+++ tests/camel-itest-karaf/src/test/java/org/apache/camel/itest/karaf/main/BundleIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.core.osgi.main.it;
+package org.apache.camel.itest.karaf.main;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
